implemented add, edit and delete functionality

+++ Models/Item.cs
@@ -4,9 +4,9 @@ namespace MyMvcApp.Models
   {
     public int Id {get;set;}
     public string Name {get; set;} = string.Empty;
-    public double Price {get; set;}
+    public decimal Price {get; set;}
 
   }
 
   
-} 
+}
